% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decathlon_individual.R
\name{dec_100m}
\alias{dec_100m}
\title{Convert decathlon 100m score in to points}
\usage{
dec_100m(X100m)
}
\arguments{
\item{X100m}{A 100m time, in seconds}
}
\description{
Convert decathlon 100m score in to points
}
\examples{
dec_100m(10.12)
}
