% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decathlon_individual.R
\name{dec_110mh}
\alias{dec_110mh}
\title{Convert decathlon 110m hurdles score in to points}
\usage{
dec_110mh(X110mh)
}
\arguments{
\item{X110mh}{A 110mh hurdles time, in seconds}
}
\description{
Convert decathlon 110m hurdles score in to points
}
\examples{
dec_110mh(13.44)
}
