% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decathlon_individual.R
\name{dec_1500m}
\alias{dec_1500m}
\title{Convert decathlon 1500m score in to points}
\usage{
dec_1500m(X1500m)
}
\arguments{
\item{X1500m}{A 1500m time, in seconds}
}
\description{
Convert decathlon 1500m score in to points
}
\examples{
dec_1500m(238.70)
}
