% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decathlon_individual.R
\name{dec_400m}
\alias{dec_400m}
\title{Convert decathlon 400m score in to points}
\usage{
dec_400m(X400m)
}
\arguments{
\item{X400m}{A 400m time, in seconds}
}
\description{
Convert decathlon 400m score in to points
}
\examples{
dec_400m(45.00)
}
