% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decathlon_individual.R
\name{dec_dt}
\alias{dec_dt}
\title{Convert decathlon discus score in to points}
\usage{
dec_dt(DT)
}
\arguments{
\item{DT}{A discus throw measurement, in metres}
}
\description{
Convert decathlon discus score in to points
}
\examples{
dec_dt(55.87)
}
