% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decathlon_individual.R
\name{dec_hj}
\alias{dec_hj}
\title{Convert decathlon high jump score in to points}
\usage{
dec_hj(HJ)
}
\arguments{
\item{HJ}{A high jump measurement, in metres}
}
\description{
Convert decathlon high jump score in to points
}
\examples{
dec_hj(2.27)
}
