% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decathlon_individual.R
\name{dec_jt}
\alias{dec_jt}
\title{Convert decathlon javelin throw score in to points}
\usage{
dec_jt(JT)
}
\arguments{
\item{JT}{A javelin throw measurement, in metres}
}
\description{
Convert decathlon javelin throw score in to points
}
\examples{
dec_jt(79.80)
}
