% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decathlon_individual.R
\name{dec_lj}
\alias{dec_lj}
\title{Convert decathlon long jump score in to points}
\usage{
dec_lj(LJ)
}
\arguments{
\item{LJ}{A long jump measurement, in metres}
}
\description{
Convert decathlon long jump score in to points
}
\examples{
dec_lj(8.23)
}
