% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decathlon_individual.R
\name{dec_pv}
\alias{dec_pv}
\title{Convert decathlon pole vault score in to points}
\usage{
dec_pv(PV)
}
\arguments{
\item{PV}{A pole vault measurement, in metres}
}
\description{
Convert decathlon pole vault score in to points
}
\examples{
dec_pv(5.76)
}
