% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decathlon_individual.R
\name{dec_sp}
\alias{dec_sp}
\title{Convert decathlon shot put score in to points}
\usage{
dec_sp(SP)
}
\arguments{
\item{SP}{A shot put measurement, in metres}
}
\description{
Convert decathlon shot put score in to points
}
\examples{
dec_sp(19.17)
}
