% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decathlon_grab.R
\name{decathlon_grab}
\alias{decathlon_grab}
\title{Load specific decathlon event data}
\usage{
decathlon_grab(filterEvent, filterYear, filterCountry, filterRank)
}
\arguments{
\item{filterEvent}{Character vector of either \code{"olympics"}, \code{"world_championships"} or \code{"gotzis"}}

\item{filterYear}{(Optional) Numeric vector filter for years}

\item{filterCountry}{(Optional) Character vector filter for country codes}

\item{filterRank}{(Optional) Numeric vector filter for ranks}
}
\description{
Load specific decathlon event data
}
\examples{
decathlon_grab("olympics", c(2000, 2008), c("EST", "USA"), 1:3)
}
