% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decathlon_list_points.R
\docType{data}
\name{decathlon_list_points}
\alias{decathlon_list_points}
\title{List object containing decathlon data points}
\format{
A list object with three tibbles (for each of the major events) each containing 16 rows:
\describe{
\item{Year}{Year of event}
\item{date}{Date of first day of the decathlon}
\item{Athlete}{Name of decathlete}
\item{Country}{IOC country code of decathlete}
\item{Rank}{Final rank of decathlete out of those who completed the event}
\item{Final Score}{Final score attained at end of the event}
\item{100m}{Numeric vector}
\item{LJ}{Numeric vector}
\item{SP}{Numeric vector}
\item{HJ}{Numeric vector}
\item{400m}{Numeric vector}
\item{11mh}{Numeric vector}
\item{DT}{Numeric vector}
\item{PV}{Numeric vector}
\item{JT}{Numeric vector}
\item{1500m}{Numeric vector}
}
}
\source{
Most of the data were admittedly initially collected from Wikipedia (during the days of where my R skills were lacking), then later cross-checked and verified from official websites including \url{http://www.euromeetings.org/history.asp?id=8}, \url{https://www.sports-reference.com}, \url{https://www.iaaf.org}, \url{https://www.decathlon2000.com/}, with additional testing done through converting raw measurements and times through own functions.
}
\usage{
decathlon_list_points
}
\description{
A list object containing tibbles of decathlete names, International Olympic Committee (IOC) country codes, final ranking and event point scores for decathletes who completed the decathlon at the Olympics, IAAF World Championships, and Götzis Hypomeeting events from 2000 onwards.
}
\keyword{datasets}
