% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decathlon_vis.R
\name{decathlon_vis}
\alias{decathlon_vis}
\title{Create visualisation of decathlon scores}
\usage{
decathlon_vis(...)
}
\arguments{
\item{...}{Arguments to \code{decathlon_s2p}}
}
\description{
\code{decathlon_vis} creates a visualisation for an individual completed decathlon.
}
