% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heptathlon_individual.R
\name{hept_100mh}
\alias{hept_100mh}
\title{Convert heptathlon 100m hurdles score in to points}
\usage{
hept_100mh(X100mh)
}
\arguments{
\item{X100mh}{A 100m hurdles time, in seconds}
}
\description{
Convert heptathlon 100m hurdles score in to points
}
\examples{
hept_100mh(12.54)
}
