% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heptathlon_individual.R
\name{hept_200m}
\alias{hept_200m}
\title{Convert heptathlon 200m score in to points}
\usage{
hept_200m(X200m)
}
\arguments{
\item{X200m}{A 200m time, in seconds}
}
\description{
Convert heptathlon 200m score in to points
}
\examples{
hept_200m(22.30)
}
