% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heptathlon_individual.R
\name{hept_800m}
\alias{hept_800m}
\title{Convert heptathlon 800m score in to points}
\usage{
hept_800m(X800m)
}
\arguments{
\item{X800m}{A 800m time, in seconds}
}
\description{
Convert heptathlon 800m score in to points
}
\examples{
hept_800m(121.84)
}
