% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heptathlon_individual.R
\name{hept_hj}
\alias{hept_hj}
\title{Convert heptathlon high jump score in to points}
\usage{
hept_hj(HJ)
}
\arguments{
\item{HJ}{A high jump height measurement, in metres}
}
\description{
Convert heptathlon high jump score in to points
}
\examples{
hept_hj(2.02)
}
