% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heptathlon_individual.R
\name{hept_jt}
\alias{hept_jt}
\title{Convert heptathlon javelin throw score in to points}
\usage{
hept_jt(JT)
}
\arguments{
\item{JT}{A javelin throw distance measurement, in metres}
}
\description{
Convert heptathlon javelin throw score in to points
}
\examples{
hept_jt(60.90)
}
