% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heptathlon_individual.R
\name{hept_lj}
\alias{hept_lj}
\title{Convert heptathlon long jump score in to points}
\usage{
hept_lj(LJ)
}
\arguments{
\item{LJ}{A long jump distance measurement, in metres}
}
\description{
Convert heptathlon long jump score in to points
}
\examples{
hept_lj(7.27)
}
