% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heptathlon_individual.R
\name{hept_sp}
\alias{hept_sp}
\title{Convert heptathlon shot put score in to points}
\usage{
hept_sp(SP)
}
\arguments{
\item{SP}{A shot put distance measurement, in metres}
}
\description{
Convert heptathlon shot put score in to points
}
\examples{
hept_sp(17.31)
}
