% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heptathlon_s2p.R
\name{heptathlon_s2p}
\alias{heptathlon_s2p}
\title{Convert heptathlon scores in to points}
\usage{
heptathlon_s2p(X100mh, HJ, SP, X200m, LJ, JT, X800m)
}
\arguments{
\item{X100mh}{A 100m hurdles time, in seconds}

\item{HJ}{A high jump measurement, in m}

\item{SP}{A shot put measurement, in m}

\item{X200m}{A 200m time, in seconds}

\item{LJ}{A long jump measurement, in m}

\item{JT}{A javelin throw measurement, in m}

\item{X800m}{A 800m time, in seconds}
}
\description{
\code{heptathlon_s2p} calculates scores for performances in the heptathlon, as
well as providing useful descriptive and summary statistics.
}
