% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heptathlon_vis.R
\name{heptathlon_vis}
\alias{heptathlon_vis}
\title{Create visualisation of heptathlon scores}
\usage{
heptathlon_vis(...)
}
\arguments{
\item{...}{Arguments to \code{heptathlon_s2p}}
}
\description{
\code{heptathlon_vis} creates a visualisation for an individual completed heptathlon.
}
