% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_calculations.R
\name{jumps_func}
\alias{jumps_func}
\title{Internal helper functions for converting jumping event scores in to points}
\usage{
jumps_func(kitty, a, b, c)
}
\arguments{
\item{kitty}{The event argument from \code{decathlon_s2p}}
}
\description{
Internal helper functions for converting jumping event scores in to points
}
\keyword{internal}
