% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_calculations.R
\name{runs_func}
\alias{runs_func}
\title{Internal helper functions for converting running event scores in to points}
\usage{
runs_func(chonk, a, b, c)
}
\arguments{
\item{chonk}{The event argument from \code{decathlon_s2p}}
}
\description{
Internal helper functions for converting running event scores in to points
}
\keyword{internal}
