% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_calculations.R
\name{throws_func}
\alias{throws_func}
\title{Internal helper functions for converting throwing event scores in to points}
\usage{
throws_func(moo, a, b, c)
}
\arguments{
\item{moo}{The event argument from \code{decathlon_s2p}}
}
\description{
Internal helper functions for converting throwing event scores in to points
}
\keyword{internal}
