% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decathlon_grab.R
\name{decathlon_grab}
\alias{decathlon_grab}
\title{Load decathlon event points data}
\usage{
decathlon_grab(filterEvent = "all", filterYear, filterCountry, filterRank)
}
\arguments{
\item{filterEvent}{Character vector of either \code{"all"}, \code{"olympics"}, \code{"world_championships"}, \code{"gotzis"}, or a combination of the major events. Defaults to \code{"all"}.}

\item{filterYear}{(Optional) Numeric vector filter for years}

\item{filterCountry}{(Optional) Character vector filter for IOC country codes. These can be found on the pertinent \href{https://en.wikipedia.org/wiki/List_of_IOC_country_codes}{Wikipedia page}, or accessed through the CRAN package \code{countrycode}.}

\item{filterRank}{(Optional) Numeric vector filter for ranks}
}
\description{
Load decathlon event points data
}
\examples{
decathlon_grab("olympics", c(2000, 2008), c("EST", "USA"), 1:3)
}
