% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decathlon_s2p.R
\name{decathlon_s2p}
\alias{decathlon_s2p}
\title{Convert decathlon scores in to points}
\usage{
decathlon_s2p(X100m, LJ, SP, HJ, X400m, X110mh, DT, PV, JT, X1500m)
}
\arguments{
\item{X100m}{A 100m time, in seconds}

\item{LJ}{A long jump measurement, in m}

\item{SP}{A shot put measurement, in m}

\item{HJ}{A high jump measurement, in m}

\item{X400m}{A 400m time, in seconds}

\item{X110mh}{A 110m hurdles time, in seconds}

\item{DT}{A discus throw measurement, in m}

\item{PV}{A pole vault measurement, in m}

\item{JT}{A javelin throw measurement, in m}

\item{X1500m}{A 1500m time, in seconds}
}
\description{
\code{decathlon_s2p} calculates scores for performances in the decathlon, as
well as providing useful descriptive and summary statistics.
}
